Load project files
This commit allows for a more robust file structure for a project. By
running build_app and run_krull from the root directory of the project,
all files in the root and all files in the recursive root directories,
will be processesed and registered to the krull app where necessary.
In practice, for this commit, this allows us to register endpoints in
multiple files. In the future this will allow us to register a multitude
of processes like error handling, middleware, etc.